--- src/main/java/hello/batch/StepTimerListener.java
@@ -1,23 +1,0 @@
-package hello.batch;
-
-import org.springframework.batch.core.ExitStatus;
-import org.springframework.batch.core.StepExecution;
-import org.springframework.batch.core.StepExecutionListener;
-
-public class StepTimerListener implements StepExecutionListener {
-
-    private long startTime;
-
-    @Override
-    public void beforeStep(StepExecution stepExecution) {
-        startTime = System.currentTimeMillis();
-    }
-
-    @Override
-    public ExitStatus afterStep(StepExecution stepExecution) {
-        long endTime = System.currentTimeMillis();
-        long duration = endTime - startTime;
-        System.out.println("Step took " + duration + " ms");
-        return stepExecution.getExitStatus();
-    }
-}
